+++ test/libsolidity/semanticTests/using/free_functions_individual.sol
@@ -1,0 +1,27 @@
+function id(uint x) pure returns (uint) {
+    return x;
+}
+
+function zero(uint) pure returns (uint) {
+    return 0;
+}
+
+contract C {
+    using id for uint;
+    using zero for uint;
+
+    function f(uint z) pure external returns(uint) {
+        return z.id();
+    }
+
+    function g(uint z) pure external returns (uint) {
+        return z.zero();
+    }
+}
+// ====
+// compileViaYul: also
+// ----
+// f(uint256): 10 -> 10
+// g(uint256): 10 -> 0
+// f(uint256): 256 -> 0x0100
+// g(uint256): 256 -> 0