+++ test/libsolidity/syntaxTests/bound/bound_to_struct.sol
@@ -7,4 +7,4 @@ contract C {
     using S for S;
 }
 // ----
-// TypeError 4357: (113-114): Library name expected.
+// TypeError 8187: (113-114): Expected library name, free-function(s) name(s), module name or *.