+++ test/libsolidity/syntaxTests/using/module_1.sol
@@ -1,0 +1,13 @@
+==== Source: A ====
+function id(uint x) pure returns (uint) {
+    return x;
+}
+
+==== Source: B ====
+import "A" as M;
+contract C {
+    using M for uint;
+	function f(uint x) public pure returns (uint) {
+        return x.id();
+    }
+}