+++ test/libsolidity/syntaxTests/using/using_for_ast_file_level.sol
@@ -1,0 +1,7 @@
+function id(uint x) pure returns (uint) {
+    return x;
+}
+
+using id for *;
+// ----
+// SyntaxError 8118: (59-74): The type has to be specified explicitly at file level (cannot use '*').