+++ test/libsolidity/syntaxTests/using/using_free_functions_1.sol
@@ -1,0 +1,11 @@
+function id(uint x) pure returns (uint) {
+    return x;
+}
+
+contract C {
+    using id for uint;
+
+    function g(uint z) pure external {
+        z.id();
+    }
+}