+++ test/libsolidity/syntaxTests/using/using_free_no_parameters_err.sol
@@ -1,0 +1,9 @@
+function one() pure returns(uint) {
+    return 1;
+}
+
+contract C {
+    using one for uint;
+}
+// ----
+// TypeError 4731: (76-79): The function "one" does not have any parameters, and therefore cannot be bound to the type "uint256".