+++ test/libsolidity/syntaxTests/using/using_non_free_function_err.sol
@@ -1,0 +1,13 @@
+contract C {
+    using {f, h} for uint;
+
+    function f(uint) internal {
+    }
+    function g(uint) external {
+    }
+    function h(uint) internal {
+    }
+}
+// ----
+// TypeError 4167: (24-25): Only free functions can be bound to a type in a "using" statement
+// TypeError 4167: (27-28): Only free functions can be bound to a type in a "using" statement